% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-funnel.R
\name{bs_model_funnel_col}
\alias{bs_model_funnel_col}
\title{Create Funnel Model from Column}
\usage{
bs_model_funnel_col(data, column, ...)
}
\arguments{
\item{data}{Data frame or sf object}

\item{column}{Name of column containing sample sizes}

\item{...}{Additional arguments passed to \code{\link[=bs_model_funnel]{bs_model_funnel()}}}
}
\value{
A \code{bs_model_funnel} object
}
\description{
Convenience function to create a funnel model from a data frame column.
}
\examples{
df <- data.frame(sample_size = c(1000, 2000, 3000))
model <- bs_model_funnel_col(df, "sample_size")
model$name
}
